continue extending the API
The PAF output is integrated now, which is OK.

This does mean that when this is used inside of another context that
might call it for small regions, we have to add in all the offsets and
lengths relative to the original sequences.

These additions to the API just expose an interface to push this into
the PAF output that's provided as a string.